+++ app/src/main/java/com/fibelatti/photowidget/configure/IntentKtx.kt
@@ -12,6 +12,8 @@ var Intent.appWidgetId: Int by intentExtras(
     default = AppWidgetManager.INVALID_APPWIDGET_ID,
 )
 
+var Intent.duplicateFromId: Int? by intentExtras()
+
 var Intent.photoWidget: PhotoWidget by intentExtras()
 
 var Intent.aspectRatio: PhotoWidgetAspectRatio by intentExtras()